Reorganized layers into new directory.
