+++ censors/censor7.py
@@ -5,7 +5,7 @@ Does not check if the SEQ/ACK are in window for the FIN/RST.
 """
 
 import logging
-import actions.packet
+import layers.packet
 logging.getLogger("scapy.runtime").setLevel(logging.ERROR)
 from scapy.all import IP, TCP
 
@@ -24,7 +24,7 @@ class Censor7(Censor):
         Check if the censor should run against this packet. Returns true or false.
         """
         try:
-            self.logger.debug("Inbound packet to censor: " + actions.packet.Packet._str_packet(packet))
+            self.logger.debug("Inbound packet to censor: " + layers.packet.Packet._str_packet(packet))
             if self.drop_all_from == packet["IP"].src:
                 self.logger.debug("Dropping all from this IP %s..." % self.drop_all_from)
                 return True
@@ -43,7 +43,7 @@ class Censor7(Censor):
                    packet["TCP"].sport in self.tcb["ports"]:
 
                     self.tcb = None
-                    self.logger.debug(("Tearing down TCB on packet " + actions.packet.Packet._str_packet(packet)))
+                    self.logger.debug(("Tearing down TCB on packet " + layers.packet.Packet._str_packet(packet)))
                     return False
 
             elif not self.tcb and self.tcb is not None:
@@ -51,13 +51,13 @@ class Censor7(Censor):
                 self.tcb["ports"] = [packet["TCP"].sport, packet["TCP"].dport]
 
             if self.tcb is None:
-                self.logger.debug("Ignoring packet: " + actions.packet.Packet._str_packet(packet))
+                self.logger.debug("Ignoring packet: " + layers.packet.Packet._str_packet(packet))
                 return False
 
             # Check if any forbidden words appear in the packet payload
             for keyword in self.forbidden:
                 if keyword in self.get_payload(packet):
-                    self.logger.debug("Packet triggered censor: " + actions.packet.Packet._str_packet(packet))
+                    self.logger.debug("Packet triggered censor: " + layers.packet.Packet._str_packet(packet))
                     return True
 
             return False
