+++ censors/censor8.py
@@ -4,7 +4,7 @@ if the full tuple of the TCB matches (src, dst, sport, dport, seq).
 """
 
 import logging
-import actions.packet
+import layers.packet
 logging.getLogger("scapy.runtime").setLevel(logging.ERROR)
 from scapy.all import IP, TCP
 
@@ -23,7 +23,7 @@ class Censor8(Censor):
         Check if the censor should run against this packet. Returns true or false.
         """
         try:
-            self.logger.debug("Inbound packet to censor: " + actions.packet.Packet._str_packet(packet))
+            self.logger.debug("Inbound packet to censor: " + layers.packet.Packet._str_packet(packet))
             if self.drop_all_from == packet["IP"].src:
                 self.logger.debug("Dropping all from this IP %s..." % self.drop_all_from)
                 return True
@@ -36,7 +36,7 @@ class Censor8(Censor):
                 self.tcb["ips"] = [packet["IP"].src, packet["IP"].dst]
                 self.tcb["ports"] = [packet["TCP"].sport, packet["TCP"].dport]
                 self.tcb["seq"] = packet["TCP"].seq + 1
-                self.logger.debug("Synchronizing TCB on packet " + actions.packet.Packet._str_packet(packet))
+                self.logger.debug("Synchronizing TCB on packet " + layers.packet.Packet._str_packet(packet))
                 return False
             # TCB teardown
             elif packet["TCP"].sprintf('%TCP.flags%') == "R" or packet["TCP"].sprintf('%TCP.flags%') == "F":
@@ -49,11 +49,11 @@ class Censor8(Censor):
                    packet["TCP"].seq == self.tcb["seq"]:
 
                     self.tcb = None
-                    self.logger.debug(("Tearing down TCB on packet " + actions.packet.Packet._str_packet(packet)))
+                    self.logger.debug(("Tearing down TCB on packet " + layers.packet.Packet._str_packet(packet)))
                     return False
 
             if self.tcb is None:
-                self.logger.debug("Ignoring packet: " + actions.packet.Packet._str_packet(packet))
+                self.logger.debug("Ignoring packet: " + layers.packet.Packet._str_packet(packet))
                 return False
             elif "seq" in self.tcb and packet["TCP"].seq == self.tcb["seq"]:
                 self.tcb["seq"] += len(self.get_payload(packet))
@@ -61,7 +61,7 @@ class Censor8(Censor):
             # Check if any forbidden words appear in the packet payload
             for keyword in self.forbidden:
                 if keyword in self.get_payload(packet):
-                    self.logger.debug("Packet triggered censor: " + actions.packet.Packet._str_packet(packet))
+                    self.logger.debug("Packet triggered censor: " + layers.packet.Packet._str_packet(packet))
                     return True
 
             return False
