+++ tests/test_duplicate.py
@@ -3,7 +3,7 @@ import sys
 sys.path.append("..")
 
 import actions.duplicate
-import actions.packet
+import layers.packet
 import actions.strategy
 import actions.utils
 import evolve
@@ -18,7 +18,7 @@ def test_duplicate(logger):
     duplicate = actions.duplicate.DuplicateAction()
     assert str(duplicate) == "duplicate", "Duplicate returned incorrect string representation: %s" % str(duplicate)
 
-    packet = actions.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
+    packet = layers.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
     packet1, packet2 = duplicate.run(packet, logger)
     assert id(packet1) != id(packet2), "Duplicate aliased packet objects"
     duplicate.mutate()
