+++ tests/test_evolve.py
@@ -12,7 +12,7 @@ import common
 import evolve
 import evaluator
 import actions.utils
-import actions.packet
+import layers.packet
 from actions.tamper import TamperAction
 from scapy.all import IP, TCP, UDP
 import random
@@ -50,7 +50,7 @@ def test_disable_single_action(logger):
      """
      Tests disabling a single action
      """
-     actions.packet.Packet.reset_restrictions()
+     layers.packet.Packet.reset_restrictions()
      try:
          logger.setLevel("ERROR")
          actions.action.ACTION_CACHE={}
@@ -66,14 +66,14 @@ def test_disable_single_action(logger):
              actions.action.ACTION_CACHE["in"] = {}
              actions.action.ACTION_CACHE["out"] = {}
      finally:
-         actions.packet.Packet.reset_restrictions()
+         layers.packet.Packet.reset_restrictions()
 
 
 def test_disable_multiple_actions(logger):
     """
     Tests disabling multiple actions
     """
-    actions.packet.Packet.reset_restrictions()
+    layers.packet.Packet.reset_restrictions()
     try:
         logger.setLevel("ERROR")
         actions.action.ACTION_CACHE={}
@@ -95,7 +95,7 @@ def test_disable_multiple_actions(logger):
             actions.action.ACTION_CACHE["in"] = {}
             actions.action.ACTION_CACHE["out"] = {}
     finally:
-        actions.packet.Packet.reset_restrictions()
+        layers.packet.Packet.reset_restrictions()
 
 
 def assert_only(ind, field):
@@ -160,7 +160,7 @@ def test_disable_fields(logger, use_canary):
                 p.mutate(logger)
                 assert_only(p, "ack")
 
-        actions.packet.Packet.reset_restrictions()
+        layers.packet.Packet.reset_restrictions()
 
         # Restrict evolve to using NOT the dataofs or chksum field in the TCP header
         evolve.restrict_headers(logger, "TCP,UDP", "", "dataofs,chksum",)
@@ -194,7 +194,7 @@ def test_disable_fields(logger, use_canary):
                 assert_not(p, ["dataofs", "chksum"])
 
     finally:
-        actions.packet.Packet.reset_restrictions()
+        layers.packet.Packet.reset_restrictions()
 
 
 @pytest.mark.parametrize("use_canary", [True, False], ids=["with_canary", "without_canary"])
@@ -220,7 +220,7 @@ def test_population_pool(logger, use_canary):
         tester = evaluator.Evaluator(cmd, logger)
         canary_id = evolve.run_collection_phase(logger, tester)
 
-    actions.packet.Packet.reset_restrictions()
+    layers.packet.Packet.reset_restrictions()
     population = []
     print("Generating population pool")
     # Generate random strategies to initialize the population
@@ -237,7 +237,7 @@ def test_population_pool(logger, use_canary):
         IP(src="127.0.0.1", dst="127.0.0.1")/UDP(sport=2222, dport=3333, chksum=0x4444),
         IP(src="127.0.0.1", dst="127.0.0.1")/UDP(sport=2222, dport=3333, chksum=0x8888)
     ]
-    packets = [actions.packet.Packet(packet) for packet in packets]
+    packets = [layers.packet.Packet(packet) for packet in packets]
     for generation in range(0, 20):
         print("Starting fake generation %d" % generation)
         for ind in population:
@@ -249,7 +249,7 @@ def test_population_pool(logger, use_canary):
                     print(str(ind))
                     print(packet)
                     packet.show()
-                    print(actions.packet.SUPPORTED_LAYERS)
+                    print(layers.packet.SUPPORTED_LAYERS)
                     raise
         for p in population:
             try:
@@ -305,7 +305,7 @@ def test_mutation(logger):
     Tests mutation.
     """
 
-    actions.packet.Packet.reset_restrictions()
+    layers.packet.Packet.reset_restrictions()
     population = [actions.utils.parse("[TCP:flags:PA]-| \/", logger)]
     population[0].in_enabled = False
     assert population
