+++ tests/test_packet.py
@@ -1,7 +1,14 @@
 import pytest
 
-import actions.packet
-import actions.layer
+import actions.utils
+import actions.trigger
+import layers.packet
+import layers.layer
+import layers.tcp_layer
+import layers.dns_layer
+import layers.dnsqr_layer
+import layers.udp_layer
+import layers.ip_layer
 import evolve
 
 from scapy.all import IP, TCP, UDP, DNS, DNSQR, Raw, DNSRR
@@ -12,10 +19,10 @@ def test_parse_layers():
     Tests layer parsing.
     """
     pkt = IP()/TCP()/Raw("")
-    packet = actions.packet.Packet(pkt)
-    layers = list(packet.read_layers())
-    assert layers[0].name == "IP"
-    assert layers[1].name == "TCP"
+    packet = layers.packet.Packet(pkt)
+    layers_l = list(packet.read_layers())
+    assert layers_l[0].name == "IP"
+    assert layers_l[1].name == "TCP"
 
     layers_dict = packet.setup_layers()
     assert layers_dict["IP"]
@@ -27,9 +34,9 @@ def test_get_random():
     Tests get random
     """
 
-    tcplayer = actions.layer.TCPLayer(TCP())
+    tcplayer = layers.tcp_layer.TCPLayer(TCP())
     field, value = tcplayer.get_random()
-    assert field in actions.layer.TCPLayer.fields
+    assert field in layers.tcp_layer.TCPLayer.fields
 
 
 def test_gen_random():
@@ -37,7 +44,7 @@ def test_gen_random():
     Tests gen random
     """
     for i in range(0, 2000):
-        layer, field, value = actions.packet.Packet().gen_random()
+        layer, field, value = layers.packet.Packet().gen_random()
         assert layer in [DNS, TCP, UDP, IP, DNSQR]
 
 
@@ -47,14 +54,14 @@ def test_dnsqr():
     """
     pkt = UDP()/DNS(ancount=1)/DNSQR()
     pkt.show()
-    packet = actions.packet.Packet(pkt)
+    packet = layers.packet.Packet(pkt)
     packet.show()
     assert len(packet.layers) == 3
     assert "UDP" in packet.layers
     assert "DNS" in packet.layers
     assert "DNSQR" in packet.layers
     pkt = IP()/UDP()/DNS()/DNSQR()
-    packet = actions.packet.Packet(pkt)
+    packet = layers.packet.Packet(pkt)
     assert str(packet)
 
 
@@ -62,12 +69,12 @@ def test_load():
     """
     Tests loads.
     """
-    tcp = actions.layer.TCPLayer(TCP())
+    tcp = layers.tcp_layer.TCPLayer(TCP())
     load = tcp.gen("load")
     pkt = IP()/"datadata"
-    p = actions.packet.Packet(pkt)
+    p = layers.packet.Packet(pkt)
     assert p.get("IP", "load") == "datadata"
-    p2 = actions.packet.Packet(IP(bytes(p)))
+    p2 = layers.packet.Packet(IP(bytes(p)))
     assert p2.get("IP", "load") == "datadata"
     p2.set("IP", "load", "data2")
     # Check p is unchanged
@@ -81,9 +88,9 @@ def test_load():
     assert p2.get("IP", "chksum") == None
 
     pkt = IP()/TCP()/"datadata"
-    p = actions.packet.Packet(pkt)
+    p = layers.packet.Packet(pkt)
     assert p.get("TCP", "load") == "datadata"
-    p2 = actions.packet.Packet(IP(bytes(p)))
+    p2 = layers.packet.Packet(IP(bytes(p)))
     assert p2.get("TCP", "load") == "datadata"
     p2.set("TCP", "load", "data2")
     # Check p is unchanged
@@ -97,7 +104,7 @@ def test_parse_load(logger):
     """
     Tests load parsing.
     """
-    pkt = actions.packet.Packet(IP()/TCP()/"TYPE A\r\n")
+    pkt = layers.packet.Packet(IP()/TCP()/"TYPE A\r\n")
     print("Parsed: %s" % pkt.get("TCP", "load"))
 
     strat = actions.utils.parse("[TCP:load:TYPE%20A%0D%0A]-drop-| \/", logger)
@@ -113,12 +120,12 @@ def test_dns():
     """
     Tests DNS layer.
     """
-    dns = actions.layer.DNSLayer(DNS())
+    dns = layers.dns_layer.DNSLayer(DNS())
     print(dns.gen("id"))
     assert dns.gen("id")
 
-    p = actions.packet.Packet(DNS(id=0xabcd))
-    p2 = actions.packet.Packet(DNS(bytes(p)))
+    p = layers.packet.Packet(DNS(id=0xabcd))
+    p2 = layers.packet.Packet(DNS(bytes(p)))
     assert p.get("DNS", "id") == 0xabcd
     assert p2.get("DNS", "id") == 0xabcd
 
@@ -126,13 +133,13 @@ def test_dns():
     assert p.get("DNS", "id") == 0xabcd # Check p is unchanged
     assert p2.get("DNS", "id") == 0x4321
 
-    dns = actions.packet.Packet(DNS(aa=1))
+    dns = layers.packet.Packet(DNS(aa=1))
     assert dns.get("DNS", "aa") == 1
     aa = dns.gen("DNS", "aa")
     assert aa == 0 or aa == 1
     assert dns.get("DNS", "aa") == 1 # Original value unchanged
 
-    dns = actions.packet.Packet(DNS(opcode=15))
+    dns = layers.packet.Packet(DNS(opcode=15))
     assert dns.get("DNS", "opcode") == 15
     opcode = dns.gen("DNS", "opcode")
     assert opcode >= 0 and opcode <= 15
@@ -141,7 +148,7 @@ def test_dns():
     dns.set("DNS", "opcode", 3)
     assert dns.get("DNS", "opcode") == 3
 
-    dns = actions.packet.Packet(DNS(qr=0))
+    dns = layers.packet.Packet(DNS(qr=0))
     assert dns.get("DNS", "qr") == 0
     qr = dns.gen("DNS", "qr")
     assert qr == 0 or qr == 1
@@ -150,7 +157,7 @@ def test_dns():
     dns.set("DNS", "qr", 1)
     assert dns.get("DNS", "qr") == 1
 
-    dns = actions.packet.Packet(DNS(arcount=0xAABB))
+    dns = layers.packet.Packet(DNS(arcount=0xAABB))
     assert dns.get("DNS", "arcount") == 0xAABB
     arcount = dns.gen("DNS", "arcount")
     assert arcount >= 0 and arcount <= 0xffff
@@ -159,13 +166,13 @@ def test_dns():
     dns.set("DNS", "arcount", 65432)
     assert dns.get("DNS", "arcount") == 65432
 
-    dns = actions.layer.DNSLayer(DNS()/DNSQR(qname="example.com"))
+    dns = layers.dns_layer.DNSLayer(DNS()/DNSQR(qname="example.com"))
     assert isinstance(dns.get_next_layer(), DNSQR)
     print(dns.gen("id"))
     assert dns.gen("id")
 
-    p = actions.packet.Packet(DNS(id=0xabcd))
-    p2 = actions.packet.Packet(DNS(bytes(p)))
+    p = layers.packet.Packet(DNS(id=0xabcd))
+    p2 = layers.packet.Packet(DNS(bytes(p)))
     assert p.get("DNS", "id") == 0xabcd
     assert p2.get("DNS", "id") == 0xabcd
 
@@ -175,27 +182,27 @@ def test_read_layers():
     Tests the ability to read each layer
     """
     packet = IP() / UDP() / TCP() / DNS() / DNSQR(qname="example.com") / DNSQR(qname="example2.com") / DNSQR(qname="example3.com")
-    packet_geneva = actions.packet.Packet(packet)
+    packet_geneva = layers.packet.Packet(packet)
     packet_geneva.setup_layers()
 
     i = 0
     for layer in packet_geneva.read_layers():
         if i == 0:
-            assert isinstance(layer, actions.layer.IPLayer)
+            assert isinstance(layer, layers.ip_layer.IPLayer)
         elif i == 1:
-            assert isinstance(layer, actions.layer.UDPLayer)
+            assert isinstance(layer, layers.udp_layer.UDPLayer)
         elif i == 2:
-            assert isinstance(layer, actions.layer.TCPLayer)
+            assert isinstance(layer, layers.tcp_layer.TCPLayer)
         elif i == 3:
-            assert isinstance(layer, actions.layer.DNSLayer)
+            assert isinstance(layer, layers.dns_layer.DNSLayer)
         elif i == 4:
-            assert isinstance(layer, actions.layer.DNSQRLayer)
+            assert isinstance(layer, layers.dnsqr_layer.DNSQRLayer)
             assert layer.layer.qname == b"example.com"
         elif i == 5:
-            assert isinstance(layer, actions.layer.DNSQRLayer)
+            assert isinstance(layer, layers.dnsqr_layer.DNSQRLayer)
             assert layer.layer.qname == b"example2.com"
         elif i == 6:
-            assert isinstance(layer, actions.layer.DNSQRLayer)
+            assert isinstance(layer, layers.dnsqr_layer.DNSQRLayer)
             assert layer.layer.qname == b"example3.com"
         i += 1
 
@@ -204,7 +211,7 @@ def test_multi_opts():
     Tests various option getting/setting.
     """
     pkt = IP()/TCP(options=[('MSS', 1460), ('SAckOK', b''), ('Timestamp', (4154603075, 0)), ('NOP', None), ('WScale', 7), ('md5header', b'abcd' * 8)])
-    packet = actions.packet.Packet(pkt)
+    packet = layers.packet.Packet(pkt)
     assert packet.get("TCP", "options-sackok") == ''
     assert packet.get("TCP", "options-mss") == 1460
     assert packet.get("TCP", "options-timestamp") == 4154603075
@@ -215,7 +222,7 @@ def test_multi_opts():
     assert packet.get("TCP", "options-timestamp") == 400000000
     assert packet.get("TCP", "options-wscale") == 7
     pkt = IP()/TCP(options=[('SAckOK', b''), ('Timestamp', (4154603075, 0)), ('NOP', None), ('WScale', 7)])
-    packet = actions.packet.Packet(pkt)
+    packet = layers.packet.Packet(pkt)
     # If the option isn't present, it will be returned as an empty string
     assert packet.get("TCP", "options-mss") == ''
     packet.set("TCP", "options-mss", "")
@@ -227,11 +234,11 @@ def test_options_eol():
     Tests options-eol.
     """
     pkt = TCP(options=[("EOL", None)])
-    p = actions.packet.Packet(pkt)
+    p = layers.packet.Packet(pkt)
     assert p.get("TCP", "options-eol") == ""
-    p2 = actions.packet.Packet(TCP(bytes(p)))
+    p2 = layers.packet.Packet(TCP(bytes(p)))
     assert p2.get("TCP", "options-eol") == ""
-    p = actions.packet.Packet(IP()/TCP(options=[]))
+    p = layers.packet.Packet(IP()/TCP(options=[]))
     assert p.get("TCP", "options-eol") == ""
     p.set("TCP", "options-eol", "")
     p.show()
@@ -249,8 +256,8 @@ def test_compression_fallback(logger):
     Test that compression does not touch a packet without DNS in it packet
     """
     pkt = UDP()
-    p = actions.packet.Packet(pkt)
-    p2 = actions.layer.DNSLayer.dns_decompress(p, logger)
+    p = layers.packet.Packet(pkt)
+    p2 = layers.dns_layer.DNSLayer.dns_decompress(p, logger)
     assert p2 == p, "dns_decompress changed a non DNS packet"
 
 
@@ -259,11 +266,11 @@ def test_options_mss():
     Tests options-eol.
     """
     pkt = TCP(options=[("MSS", 1440)])
-    p = actions.packet.Packet(pkt)
+    p = layers.packet.Packet(pkt)
     assert p.get("TCP", "options-mss") == 1440
-    p2 = actions.packet.Packet(TCP(bytes(p)))
+    p2 = layers.packet.Packet(TCP(bytes(p)))
     assert p2.get("TCP", "options-mss") == 1440
-    p = actions.packet.Packet(TCP(options=[]))
+    p = layers.packet.Packet(TCP(options=[]))
     assert p.get("TCP", "options-mss") == ""
     p.set("TCP", "options-mss", 2880)
     p.show()
@@ -281,7 +288,7 @@ def check_get(protocol, field, value):
     """
     pkt = protocol()
     setattr(pkt, field, value)
-    packet = actions.packet.Packet(pkt)
+    packet = layers.packet.Packet(pkt)
     assert packet.get(protocol.__name__, field) == value
 
 
@@ -374,11 +381,11 @@ def check_set_get(protocol, field, value):
     """
     Checks if the get method worked for this protocol, field, and value.
     """
-    pkt = actions.packet.Packet(protocol())
+    pkt = layers.packet.Packet(protocol())
     pkt.set(protocol.__name__, field, value)
     assert pkt.get(protocol.__name__, field) == value
     # Rebuild the packet to confirm the type survived
-    pkt2 = actions.packet.Packet(protocol(bytes(pkt)))
+    pkt2 = layers.packet.Packet(protocol(bytes(pkt)))
     assert pkt2.get(protocol.__name__, field) == value, "Value %s for header %s didn't survive packet parsing." % (value, field)
 
 
@@ -396,12 +403,12 @@ def check_gen_set_get(protocol, field):
     """
     Checks if the get method worked for this protocol, field, and value.
     """
-    pkt = actions.packet.Packet(protocol())
+    pkt = layers.packet.Packet(protocol())
     new_value = pkt.gen(protocol.__name__, field)
     pkt.set(protocol.__name__, field, new_value)
     assert pkt.get(protocol.__name__, field) == new_value
     # Rebuild the packet to confirm the type survived
-    pkt2 = actions.packet.Packet(protocol(bytes(pkt)))
+    pkt2 = layers.packet.Packet(protocol(bytes(pkt)))
     assert pkt2.get(protocol.__name__, field) == new_value
 
 
@@ -422,7 +429,7 @@ def test_custom_get():
     Tests value retrieval for custom getters.
     """
     pkt = IP()/TCP()/Raw(load="AAAA")
-    tcp = actions.packet.Packet(pkt)
+    tcp = layers.packet.Packet(pkt)
     assert tcp.get("TCP", "load") == "AAAA"
 
 
@@ -430,51 +437,51 @@ def test_restrict_fields(logger):
     """
     Tests packet field restriction.
     """
-    actions.packet.SUPPORTED_LAYERS = [
-        actions.layer.IPLayer,
-        actions.layer.TCPLayer,
-        actions.layer.UDPLayer
+    layers.packet.SUPPORTED_LAYERS = [
+        layers.ip_layer.IPLayer,
+        layers.tcp_layer.TCPLayer,
+        layers.udp_layer.UDPLayer
     ]
-    tcpfields = actions.layer.TCPLayer.fields
-    udpfields = actions.layer.UDPLayer.fields
-    ipfields = actions.layer.IPLayer.fields
+    tcpfields = layers.tcp_layer.TCPLayer.fields
+    udpfields = layers.udp_layer.UDPLayer.fields
+    ipfields = layers.ip_layer.IPLayer.fields
 
-    actions.packet.Packet.restrict_fields(logger, ["TCP", "UDP"], [], [])
-    assert len(actions.packet.SUPPORTED_LAYERS) == 2
-    assert actions.layer.TCPLayer in actions.packet.SUPPORTED_LAYERS
-    assert actions.layer.UDPLayer in actions.packet.SUPPORTED_LAYERS
-    assert not actions.layer.IPLayer in actions.packet.SUPPORTED_LAYERS
+    layers.packet.Packet.restrict_fields(logger, ["TCP", "UDP"], [], [])
+    assert len(layers.packet.SUPPORTED_LAYERS) == 2
+    assert layers.tcp_layer.TCPLayer in layers.packet.SUPPORTED_LAYERS
+    assert layers.udp_layer.UDPLayer in layers.packet.SUPPORTED_LAYERS
+    assert not layers.ip_layer.IPLayer in layers.packet.SUPPORTED_LAYERS
 
     pkt = IP()/TCP()
-    packet = actions.packet.Packet(pkt)
+    packet = layers.packet.Packet(pkt)
     assert "TCP" in packet.layers
     assert not "IP" in packet.layers
     assert len(packet.layers) == 1
 
     for i in range(0, 2000):
-        layer, proto, field = actions.packet.Packet().gen_random()
+        layer, proto, field = layers.packet.Packet().gen_random()
         assert layer in [TCP, UDP]
 
     # Check we can't retrieve any IP fields
-    for field in actions.layer.IPLayer.fields:
+    for field in layers.ip_layer.IPLayer.fields:
         with pytest.raises(AssertionError):
             packet.get("IP", field)
 
     # Check we can get all the TCP fields
-    for field in actions.layer.TCPLayer.fields:
+    for field in layers.tcp_layer.TCPLayer.fields:
         packet.get("TCP", field)
 
-    actions.packet.Packet.restrict_fields(logger, ["TCP", "UDP"], ["flags"], [])
-    packet = actions.packet.Packet(pkt)
-    assert len(actions.packet.SUPPORTED_LAYERS) == 1
-    assert actions.layer.TCPLayer in actions.packet.SUPPORTED_LAYERS
-    assert not actions.layer.UDPLayer in actions.packet.SUPPORTED_LAYERS
-    assert not actions.layer.IPLayer in actions.packet.SUPPORTED_LAYERS
-    assert actions.layer.TCPLayer.fields == ["flags"]
-    assert not actions.layer.UDPLayer.fields
+    layers.packet.Packet.restrict_fields(logger, ["TCP", "UDP"], ["flags"], [])
+    packet = layers.packet.Packet(pkt)
+    assert len(layers.packet.SUPPORTED_LAYERS) == 1
+    assert layers.tcp_layer.TCPLayer in layers.packet.SUPPORTED_LAYERS
+    assert not layers.udp_layer.UDPLayer in layers.packet.SUPPORTED_LAYERS
+    assert not layers.ip_layer.IPLayer in layers.packet.SUPPORTED_LAYERS
+    assert layers.tcp_layer.TCPLayer.fields == ["flags"]
+    assert not layers.udp_layer.UDPLayer.fields
 
     # Check we can't retrieve any IP fields
-    for field in actions.layer.IPLayer.fields:
+    for field in layers.ip_layer.IPLayer.fields:
         with pytest.raises(AssertionError):
             packet.get("IP", field)
 
@@ -487,37 +494,37 @@ def test_restrict_fields(logger):
                 packet.get("TCP", field)
 
     for i in range(0, 2000):
-        layer, field, value = actions.packet.Packet().gen_random()
+        layer, field, value = layers.packet.Packet().gen_random()
         assert layer == TCP
         assert field == "flags"
 
         _, proto, field, value, _ = actions.trigger.Trigger.get_rand_trigger(None, 0)
         assert proto == 'TCP'
         assert field == "flags"
-    actions.packet.Packet.reset_restrictions()
-    actions.packet.SUPPORTED_LAYERS = [
-        actions.layer.IPLayer,
-        actions.layer.TCPLayer,
-        actions.layer.UDPLayer
+    layers.packet.Packet.reset_restrictions()
+    layers.packet.SUPPORTED_LAYERS = [
+        layers.ip_layer.IPLayer,
+        layers.tcp_layer.TCPLayer,
+        layers.udp_layer.UDPLayer
     ]
 
     with pytest.raises(AssertionError):
-        actions.packet.Packet.restrict_fields(logger, ["TCP", "IP"], ["notathing"], ["notathing"])
-    actions.packet.Packet.reset_restrictions()
+        layers.packet.Packet.restrict_fields(logger, ["TCP", "IP"], ["notathing"], ["notathing"])
+    layers.packet.Packet.reset_restrictions()
 
-    actions.packet.Packet.restrict_fields(logger, ["TCP", "IP"], [], ["sport", "dport", "seq", "src"])
-    packet = actions.packet.Packet(pkt)
+    layers.packet.Packet.restrict_fields(logger, ["TCP", "IP"], [], ["sport", "dport", "seq", "src"])
+    packet = layers.packet.Packet(pkt)
     packet = packet.copy()
     assert packet.has_supported_layers()
-    assert len(actions.packet.SUPPORTED_LAYERS) == 2
-    assert actions.layer.TCPLayer in actions.packet.SUPPORTED_LAYERS
-    assert not actions.layer.UDPLayer in actions.packet.SUPPORTED_LAYERS
-    assert actions.layer.IPLayer in actions.packet.SUPPORTED_LAYERS
-    assert set(actions.layer.TCPLayer.fields) == set([f for f in tcpfields if f not in ["sport", "dport", "seq"]])
-    assert set(actions.layer.IPLayer.fields) == set([f for f in ipfields if f not in ["src"]])
+    assert len(layers.packet.SUPPORTED_LAYERS) == 2
+    assert layers.tcp_layer.TCPLayer in layers.packet.SUPPORTED_LAYERS
+    assert not layers.udp_layer.UDPLayer in layers.packet.SUPPORTED_LAYERS
+    assert layers.ip_layer.IPLayer in layers.packet.SUPPORTED_LAYERS
+    assert set(layers.tcp_layer.TCPLayer.fields) == set([f for f in tcpfields if f not in ["sport", "dport", "seq"]])
+    assert set(layers.ip_layer.IPLayer.fields) == set([f for f in ipfields if f not in ["src"]])
 
     # Check we can't retrieve any IP fields
-    for field in actions.layer.IPLayer.fields:
+    for field in layers.ip_layer.IPLayer.fields:
         if field == "src":
             with pytest.raises(AssertionError):
                 packet.get("IP", field)
@@ -533,7 +540,7 @@ def test_restrict_fields(logger):
             packet.get("TCP", field)
 
     for i in range(0, 2000):
-        layer, field, value = actions.packet.Packet().gen_random()
+        layer, field, value = layers.packet.Packet().gen_random()
         assert layer in [TCP, IP]
         assert field not in ["sport", "dport", "seq", "src"]
 
@@ -541,24 +548,24 @@ def test_restrict_fields(logger):
         assert proto in ['TCP', 'IP']
         assert field not in ["sport", "dport", "seq", "src"]
 
-    actions.packet.Packet.reset_restrictions()
-    actions.packet.SUPPORTED_LAYERS = [
-        actions.layer.IPLayer,
-        actions.layer.TCPLayer,
-        actions.layer.UDPLayer
+    layers.packet.Packet.reset_restrictions()
+    layers.packet.SUPPORTED_LAYERS = [
+        layers.ip_layer.IPLayer,
+        layers.tcp_layer.TCPLayer,
+        layers.udp_layer.UDPLayer
     ]
 
     evolve.restrict_headers(logger, "ip,udp,dns", "", "version")
-    packet = actions.packet.Packet(pkt)
+    packet = layers.packet.Packet(pkt)
     proto, field, value = packet.get_random()
     assert proto.__name__ in ["IP", "UDP"]
-    assert len(actions.packet.SUPPORTED_LAYERS) == 2
-    assert not actions.layer.TCPLayer in actions.packet.SUPPORTED_LAYERS
-    assert actions.layer.UDPLayer in actions.packet.SUPPORTED_LAYERS
-    assert actions.layer.IPLayer in actions.packet.SUPPORTED_LAYERS
-    assert set(actions.layer.IPLayer.fields) == set([f for f in ipfields if f not in ["version"]])
-    assert set(actions.layer.UDPLayer.fields) == set(udpfields)
+    assert len(layers.packet.SUPPORTED_LAYERS) == 2
+    assert not layers.tcp_layer.TCPLayer in layers.packet.SUPPORTED_LAYERS
+    assert layers.udp_layer.UDPLayer in layers.packet.SUPPORTED_LAYERS
+    assert layers.ip_layer.IPLayer in layers.packet.SUPPORTED_LAYERS
+    assert set(layers.ip_layer.IPLayer.fields) == set([f for f in ipfields if f not in ["version"]])
+    assert set(layers.udp_layer.UDPLayer.fields) == set(udpfields)
 
-    actions.packet.Packet.reset_restrictions()
-    for layer in actions.packet.SUPPORTED_LAYERS:
+    layers.packet.Packet.reset_restrictions()
+    for layer in layers.packet.SUPPORTED_LAYERS:
         assert layer.fields, '%s has no fields - reset failed!' % str(layer)
