+++ tests/test_population_files.py
@@ -4,6 +4,7 @@ sys.path.append("..") # Include the root of the project
 import evolve
 import os
 import actions.utils
+import layers.layer
 
 # Test Files Directory Setup
 test_files_directory = os.path.join("test_files")
@@ -97,7 +98,7 @@ def test_evolve_load_generation(logger):
     """
 
     generations = 2
-    actions.packet.Packet.reset_restrictions()
+    layers.packet.Packet.reset_restrictions()
 
     options = {}
     options["population_size"] = 2
