+++ tests/test_sleep.py
@@ -2,7 +2,7 @@ from scapy.all import IP, TCP
 import evolve
 import actions.utils
 import actions.strategy
-import actions.packet
+import layers.packet
 import actions.sleep
 import sys
 # Include the root of the project
@@ -16,7 +16,7 @@ def test_basic_sleep(logger):
     sleep = actions.sleep.SleepAction(.5)
     assert str(sleep) == "sleep{0.5}", "Sleep returned incorrect string representation: %s" % str(sleep)
 
-    packet = actions.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP()/("data"))
+    packet = layers.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP()/("data"))
     packet1, packet2 = sleep.run(packet, logger)
 
     assert packet1.sleep == .5, "Packet had wrong sleep value"
