+++ tests/test_tamper.py
@@ -8,10 +8,11 @@ sys.path.append("..")
 import evolve
 import evaluator
 import actions.strategy
-import actions.packet
+import layers.packet
 import actions.utils
 import actions.tamper
-import actions.layer
+import layers.layer
+import layers.ip_layer
 
 from scapy.all import IP, TCP, UDP, DNS, DNSQR, sr1
 
@@ -20,7 +21,7 @@ def test_tamper(logger):
     """
     Tests tampering with replace
     """
-    packet = actions.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
+    packet = layers.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
     original = copy.deepcopy(packet)
     tamper = actions.tamper.TamperAction(None, field="flags", tamper_type="replace", tamper_value="R")
     lpacket, rpacket = tamper.run(packet, logger)
@@ -48,7 +49,7 @@ def test_tamper_ip(logger):
     """
     Tests tampering with IP
     """
-    packet = actions.packet.Packet(IP(src='127.0.0.1', dst='127.0.0.1')/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
+    packet = layers.packet.Packet(IP(src='127.0.0.1', dst='127.0.0.1')/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
     original = copy.deepcopy(packet)
     tamper = actions.tamper.TamperAction(None, field="src", tamper_type="replace", tamper_value="192.168.1.1", tamper_proto="IP")
     lpacket, rpacket = tamper.run(packet, logger)
@@ -70,7 +71,7 @@ def test_tamper_udp(logger):
     """
     Tests tampering with UDP
     """
-    packet = actions.packet.Packet(IP(src='127.0.0.1', dst='127.0.0.1')/UDP(sport=2222, dport=53))
+    packet = layers.packet.Packet(IP(src='127.0.0.1', dst='127.0.0.1')/UDP(sport=2222, dport=53))
     original = copy.deepcopy(packet)
     tamper = actions.tamper.TamperAction(None, field="chksum", tamper_type="replace", tamper_value=4444, tamper_proto="UDP")
     lpacket, rpacket = tamper.run(packet, logger)
@@ -93,7 +94,7 @@ def test_tamper_ip_ident(logger):
     Tests tampering with IP and that the checksum is correctly changed
     """
 
-    packet = actions.packet.Packet(IP(src='127.0.0.1', dst='127.0.0.1')/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
+    packet = layers.packet.Packet(IP(src='127.0.0.1', dst='127.0.0.1')/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
     original = copy.deepcopy(packet)
     tamper = actions.tamper.TamperAction(None, field='id', tamper_type='replace', tamper_value=3333, tamper_proto="IP")
     lpacket, rpacket = tamper.run(packet, logger)
@@ -161,9 +162,9 @@ def test_mutate(logger, use_canary):
                 assert tamper.tamper_value == val, "Tamper value is not stable."
             # Create a test packet to ensure the field/proto choice was safe
             if random.random() < 0.5:
-                test_packet = actions.packet.Packet(IP()/TCP())
+                test_packet = layers.packet.Packet(IP()/TCP())
             else:
-                test_packet = actions.packet.Packet(IP()/UDP())
+                test_packet = layers.packet.Packet(IP()/UDP())
 
             # Check that tamper can run safely after mutation
             try:
@@ -199,7 +200,7 @@ def test_corrupt(logger):
     assert tamper.tamper_type == "corrupt", "Tamper action changed types."
     assert str(tamper) == "tamper{TCP:flags:corrupt}", "Tamper returned incorrect string representation: %s" % str(tamper)
 
-    packet = actions.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
+    packet = layers.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
     original = copy.deepcopy(packet)
     tamper.tamper(packet, logger)
 
@@ -226,7 +227,7 @@ def test_add(logger):
     assert tamper.tamper_type == "add", "Tamper action changed types."
     assert str(tamper) == "tamper{TCP:seq:add:10}", "Tamper returned incorrect string representation: %s" % str(tamper)
 
-    packet = actions.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
+    packet = layers.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
     original = copy.deepcopy(packet)
     tamper.tamper(packet, logger)
 
@@ -254,7 +255,7 @@ def test_decompress(logger):
     assert tamper.tamper_type == "compress", "Tamper action changed types."
     assert str(tamper) == "tamper{DNS:qd:compress}", "Tamper returned incorrect string representation: %s" % str(tamper)
 
-    packet = actions.packet.Packet(IP(dst="8.8.8.8")/UDP(dport=53)/DNS(qd=DNSQR(qname="minghui.ca.")))
+    packet = layers.packet.Packet(IP(dst="8.8.8.8")/UDP(dport=53)/DNS(qd=DNSQR(qname="minghui.ca.")))
     original = packet.copy()
     tamper.tamper(packet, logger)
     assert bytes(packet["DNS"]) == b'\x00\x00\x01\x00\x00\x02\x00\x00\x00\x00\x00\x00\x07minghui\xc0\x1a\x00\x01\x00\x01\x02ca\x00\x00\x01\x00\x01'
@@ -266,7 +267,7 @@ def test_decompress(logger):
     assert confirm_unchanged(packet, original, IP, ["len"])
     print(resp.summary())
 
-    packet = actions.packet.Packet(IP(dst="8.8.8.8")/UDP(dport=53)/DNS(qd=DNSQR(qname="maps.google.com")))
+    packet = layers.packet.Packet(IP(dst="8.8.8.8")/UDP(dport=53)/DNS(qd=DNSQR(qname="maps.google.com")))
     original = packet.copy()
     tamper.tamper(packet, logger)
     assert bytes(packet["DNS"]) == b'\x00\x00\x01\x00\x00\x02\x00\x00\x00\x00\x00\x00\x04maps\xc0\x17\x00\x01\x00\x01\x06google\x03com\x00\x00\x01\x00\x01'
@@ -279,7 +280,7 @@ def test_decompress(logger):
     print(resp.summary())
 
     # Confirm this is a NOP on normal packets
-    packet = actions.packet.Packet(IP()/UDP())
+    packet = layers.packet.Packet(IP()/UDP())
     original = packet.copy()
     tamper.tamper(packet, logger)
     assert packet.packet.summary() == original.packet.summary()
@@ -302,7 +303,7 @@ def test_corrupt_chksum(logger):
     assert tamper.tamper_type == "corrupt", "Tamper action changed types."
     assert str(tamper) == "tamper{TCP:chksum:corrupt}", "Tamper returned incorrect string representation: %s" % str(tamper)
 
-    packet = actions.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
+    packet = layers.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
     original = copy.deepcopy(packet)
     tamper.tamper(packet, logger)
 
@@ -326,7 +327,7 @@ def test_corrupt_dataofs(logger):
     """
     Tests the tamper 'replace' primitive.
     """
-    packet = actions.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S", dataofs="6L"))
+    packet = layers.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S", dataofs="6L"))
     original = copy.deepcopy(packet)
     tamper = actions.tamper.TamperAction(None, field="dataofs", tamper_type="corrupt")
 
@@ -357,7 +358,7 @@ def test_replace(logger):
     assert tamper.field == "flags", "Tamper action changed fields."
     assert tamper.tamper_type == "replace", "Tamper action changed types."
 
-    packet = actions.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
+    packet = layers.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
     original = copy.deepcopy(packet)
     tamper.tamper(packet, logger)
 
@@ -400,7 +401,7 @@ def test_parse_flags(logger):
     assert tamper.tamper_type == "replace", "Tamper action changed types."
     assert str(tamper) == "tamper{TCP:flags:replace:FRAPUN}", "Tamper returned incorrect string representation: %s" % str(tamper)
 
-    packet = actions.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
+    packet = layers.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
     tamper.tamper(packet, logger)
     assert packet[TCP].flags == "FRAPUN", "Tamper failed to change flags."
 
@@ -417,21 +418,21 @@ def test_options(logger, value, test_type):
         tamper = actions.tamper.TamperAction(None)
         assert tamper.parse("TCP:options-%s:corrupt" % value.lower(), logger)
 
-    packet = actions.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
+    packet = layers.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
     tamper.run(packet, logger)
     opts_dict_lookup = value.lower().replace(" ", "_")
 
     for optname, optval in packet["TCP"].options:
         if optname == value:
             break
-        elif optname == actions.layer.TCPLayer.options_names[opts_dict_lookup]:
+        elif optname == layers.ip_layer.TCPLayer.options_names[opts_dict_lookup]:
             break
     else:
         pytest.fail("Failed to find %s in options" % value)
     assert len(packet["TCP"].options) == 1
     raw_p = bytes(packet)
     assert raw_p, "options broke scapy bytes"
-    p2 = actions.packet.Packet(IP(bytes(raw_p)))
+    p2 = layers.packet.Packet(IP(bytes(raw_p)))
     assert p2.haslayer("IP")
     assert p2.haslayer("TCP")
     # EOLs might be added for padding, so just check >= 1
@@ -439,7 +440,7 @@ def test_options(logger, value, test_type):
     for optname, optval in p2["TCP"].options:
         if optname == value:
             break
-        elif optname == actions.layer.TCPLayer.options_names[opts_dict_lookup]:
+        elif optname == layers.ip_layer.TCPLayer.options_names[opts_dict_lookup]:
             break
     else:
         pytest.fail("Failed to find %s in options" % value)
@@ -458,7 +459,7 @@ def test_tamper_mutate_compress(logger):
         assert tamper.tamper_type == "compress"
         assert tamper.tamper_proto_str == "DNS"
         assert tamper.field == "qd"
-        packet = actions.packet.Packet(IP()/TCP()/DNS()/DNSQR())
+        packet = layers.packet.Packet(IP()/TCP()/DNS()/DNSQR())
         packet2 = tamper.tamper(packet, logger)
         assert packet2 == packet
     finally:
