+++ tests/test_trace.py
@@ -4,7 +4,7 @@ import pytest
 sys.path.append("..")
 
 import actions.trace
-import actions.packet
+import layers.packet
 import actions.strategy
 import actions.utils
 import evolve
@@ -19,14 +19,14 @@ def test_trace(logger):
     trace = actions.trace.TraceAction(start_ttl=1, end_ttl=3)
 
     assert str(trace) == "trace{1:3}", "Trace returned incorrect string representation: %s" % str(trace)
-    packet = actions.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
+    packet = layers.packet.Packet(IP(src="127.0.0.1", dst="127.0.0.1")/TCP(sport=2222, dport=3333, seq=100, ack=100, flags="S"))
     trace.run(packet, logger)
 
     print("Testing that trace will not run twice:")
     assert trace.run(packet, logger) == (None, None)
 
     trace = actions.trace.TraceAction(start_ttl=1, end_ttl=3)
-    packet = actions.packet.Packet(TCP())
+    packet = layers.packet.Packet(TCP())
     assert trace.run(packet, logger) == (packet, None)
 
     s = "[TCP:flags:PA]-trace{1:3}-| \/ "
